+++ lib/zap_client.dart
@@ -1,9 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library zap_client;
-
-export 'src/zap_socket_client.dart';
-export 'src/models/socket_event.dart' show SocketEvent;
-
-// TODO: Export any libraries intended for clients of this package.