+++ lib/zaptools_client.dart
@@ -1,0 +1,3 @@
+library zaptools_client;
+
+export 'src/client/consumer/client_connector.dart';