+++ test/zaptool_server_data.mock.dart
@@ -1,0 +1,23 @@
+
+
+
+List<String> mockData = [
+  '{"eventName":"event1", "payload": "payload1", "headers":{"myHeader":"header1"}}',
+  '{"eventName":"event2", "payload": "payload2", "headers":{"myHeader":"header2"}}',
+  '{"eventName":"event3", "payload": "payload3", "headers":{"myHeader":"header3"}}',
+];
+
+
+List<String> mockDataEvent = [
+  '{"eventName":"event1", "payload": "payload1", "headers":{"myHeader":"header1"}}',
+  '{"eventName":"event2", "payload": "payload2", "headers":{"myHeader":"header2"}}',
+  '{"eventName":"event3", "payload": "payload3", "headers":{"myHeader":"header3"}}',
+  '{"eventName":"event1", "payload": "payload1", "headers":{"myHeader":"header1"}}',
+  '{"eventName":"event2", "payload": "payload2", "headers":{"myHeader":"header2"}}',
+  '{"eventName":"event1", "payload": "payload1", "headers":{"myHeader":"header1"}}',
+  '{"eventName":"event3", "payload": "payload3", "headers":{"myHeader":"header3"}}',
+  '{"eventName":"event3", "payload": "payload3", "headers":{"myHeader":"header3"}}',
+  '{"eventName":"event3", "payload": "payload3", "headers":{"myHeader":"header3"}}',
+];
+
+